refactor(binding): merge connectable-binding into binding
